pyproject.toml
[tool.poetry]
name = "dteam-django-practical-test"
version = "0.1.0"
description = ""
authors = ["AlexAzalor <xy55xy.azalor@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.5"
django-extensions = "^4.1"
pdfkit = "^1.0.0"
djangorestframework = "^3.16.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
